test: migrate useSearchActionState tests in source
